requirements-pip.txt
###### Install by pip install -r requirements.txt #####

###### Requirements with Version Specifiers ######
docx2txt == 0.8
python-docx == 0.8.10
peewee == 3.14.4
dependency_injector == 4.31.2
pyyaml == 5.4.1
psycopg2-binary
#torch==1.7.0+cpu 
#torchvision==0.8.1+cpu 
#torchaudio===0.7.0 -f https://download.pytorch.org/whl/torch_stable.html

### Torch for linux machine 
# torch===1.7.0 
# torchvision===0.8.1 
# torchaudio===0.7.0 -f https://download.pytorch.org/whl/torch_stable.html

### Docker installation ####
torch==1.8.1+cu111
torchvision==0.9.1+cu111
torchaudio==0.8.1 

###### Requirements without Version Specifiers ######
sentence-transformers
pandas
flask 
flask-cors
PyYAML